fix(TextInput): Don't set component props to fields on ViewManager
A setting on a single TextInput has the potential to affect all other TextInput's `onSelectionChange` behavior. This makes the setting local to a single <TextInput/> component.
